Move the unit test stub CanHardware into its own module
To allow more than one unit test that accesses the
CAN hardware move the stub implementation into its
own module.

Tests:
 - Run the updated unit tests
